requirements.txt
openai>=1.0.0        # OpenRouter는 OpenAI 호환 API이므로 openai 클라이언트 사용
pyyaml>=6.0          # 실험 설정 파일(YAML) 로드
pandas>=2.0.0        # 결과 데이터 정리 및 분석
tqdm>=4.65.0         # 실험 진행 상황 표시
python-dotenv>=1.0.0 # .env 파일에서 API 키 로드
